==> final_submission/30_Driver_Attributes.Rmd <==
---
title: "30_Driver_Attributes"
output: html_notebook
---

#Import full data
```{r import}
library(tidyverse)
library(scales)

full <- read_csv("data/full.csv", col_types = cols(timestamp = col_datetime(format = "%Y-%m-%d%.%H:%M:%S%.")))
full

driver_attributes <- read_csv("data/driver_attributes.csv")
driver_attributes
```


# Creating Attributes

## Average Earnings per ride
```{r}
average_earnings <- full %>% 
  group_by(driver_id) %>% 
  select(driver_id, ride_id,driver_earnings) %>% 
  distinct() %>% 
  summarise(average_earnings_per_ride = mean(driver_earnings)) %>% 
  ungroup()

average_earnings

driver_attributes <- full_join(driver_attributes, average_earnings)
driver_attributes
  
```

## Average Ride Duration

```{r}
average_duration <- full %>% 
  group_by(driver_id) %>% 
  select(driver_id, ride_id,ride_duration) %>% 
  distinct() %>% 
  summarise(average_ride_duration = mean(ride_duration)) %>% 
  ungroup()

average_duration

driver_attributes <- full_join(driver_attributes, average_duration)
driver_attributes
  
```

## Average Ride Distance

```{r}
average_distance <- full %>% 
  group_by(driver_id) %>% 
  select(driver_id, ride_id,ride_distance) %>% 
  distinct() %>% 
  summarise(average_ride_distance = mean(ride_distance)) %>% 
  ungroup()

average_distance

driver_attributes <- full_join(driver_attributes, average_distance)
driver_attributes
  
```

## Number of rides
```{r}
number_of_rides <- full %>% 
  group_by(driver_id) %>% 
  select(driver_id, ride_id) %>% 
  distinct() %>% 
  summarise(number_of_rides_given = n()) %>% 
  ungroup()

number_of_rides

driver_attributes <- full_join(driver_attributes, number_of_rides)
driver_attributes
```
## Eagerness
Eagerness is inversely proportional to the time difference between then a ride is requested and when the driver accepts the ride

```{r}

eagerness <- full %>% 
  filter(event == "requested_at" | event == "accepted_at") %>% 
  group_by(ride_id) %>% 
  mutate(eagerness = (as.numeric(last(timestamp)-first(timestamp)))) %>% 
  ungroup() %>% 
  select(driver_id,ride_id,eagerness) %>% 
  distinct() %>% 
  group_by(driver_id) %>% 
  summarise(average_eagerness = mean(eagerness)) %>% 
  ungroup() %>% 
  mutate(scaled_average_eagerness = scales::rescale(average_eagerness,to=c(1,0))) %>% #inverse scaling
  select(driver_id,scaled_average_eagerness)
  
eagerness

driver_attributes <- full_join(driver_attributes, eagerness)
driver_attributes
```

## Response Time

Response time ranking is inversely related to time difference between a driver accepting a ride and arriving at the pickup location. This indicates the ability of a driver to choose routes where he/she does not have to drive far to get the ride.

```{r}
library(scales)
response_time <- full %>% 
  filter(event == "accepted_at" | event == "arrived_at") %>% 
  group_by(ride_id) %>% 
  mutate(response_time = (as.numeric(last(timestamp)-first(timestamp)))) %>% 
  ungroup() %>% 
  select(driver_id,ride_id,response_time) %>% 
  distinct() %>% 
  group_by(driver_id) %>% 
  summarise(average_response_time = mean(response_time)) %>% 
  ungroup() %>% 
  mutate(scaled_average_response_time_ranking = scales::rescale(average_response_time,to=c(1,0))) %>% #inverse scaling
  select(driver_id,scaled_average_response_time_ranking)
  
response_time

driver_attributes <- full_join(driver_attributes, response_time)
driver_attributes
```

## Average Prime time weighted by ride duration

```{r}
average_prime <- full %>% 
  group_by(driver_id) %>% 
  select(driver_id, ride_id, ride_duration, ride_prime_time) %>% 
  distinct() %>% 
  summarise(average_prime_time_weighted_by_ride_duration = weighted.mean(ride_prime_time,ride_duration)) %>% 
  select(driver_id,average_prime_time_weighted_by_ride_duration)

average_prime

driver_attributes <- full_join(driver_attributes, average_prime)
driver_attributes
```

## Number of Days Worked
```{r}
number_of_days_worked <- full %>% 
  mutate(day = format(as.Date(timestamp,format="%Y-%m-%d"), format = "%d")) %>% 
  mutate(month = format(as.Date(timestamp,format="%Y-%m-%d"), format = "%m")) %>% 
  group_by(driver_id) %>% 
  select(driver_id,day,month) %>% 
  distinct() %>% 
  mutate(number_of_days_worked = n()) %>% 
  select(driver_id,number_of_days_worked) %>% 
  distinct()

driver_attributes <- full_join(driver_attributes, number_of_days_worked)
driver_attributes
```


## Rides per day worked
```{r}
driver_attributes <- driver_attributes %>% 
  mutate(rides_per_day_worked = number_of_rides_given/number_of_days_worked)
driver_attributes
```


# Export driver attributes
```{r}
write_csv(driver_attributes, path = "data/driver_attributes_complete.csv")
```


# Descriptive Statistics

```{r}
average_stats <- driver_attributes %>% 
  summarise(
    `Average Earnings per Ride (USD)` = round(mean(average_earnings_per_ride), 2),
    
    `Average Ride Distance (miles)` = round(mean(average_ride_distance) * 0.000621371, 2),
    
    `Average Ride Duration (mins)` = round(mean(average_ride_duration) / 60 , 2),
    
    `Average Eagerness` = round(mean(scaled_average_eagerness), 2),
    
    `Average Responsiveness` = round(mean(scaled_average_response_time_ranking), 2),
    
    `Average Number of Rides Given` = round(mean(number_of_rides_given), 2),
    
    `Average Prime Time Weighted by Ride Duration` = round(mean(average_prime_time_weighted_by_ride_duration), 2),
    
    `Size`= n(),
    
    `Number of Days Worked` = round(mean(number_of_days_worked),2),
    
    `Average Number of Rides Per Day Worked` = round(mean(rides_per_day_worked),2)
  ) 

average_stats

average_stats %>% 
  write_csv(path = "data/average_driver_stats.csv")
```

# density plots of number of rides
```{r}
density <- driver_attributes %>% 
  select(number_of_rides_given)

density(density$number_of_rides_given)
max <- density(density$number_of_rides_given)$x[(which.max(density(density$number_of_rides_given)$y))]
second_max <- density(density$number_of_rides_given)$x[which(density(density$number_of_rides_given)$y == max(density(density$number_of_rides_given)$y[density(density$number_of_rides_given)$x > 200]))]

density %>% 
  ggplot(mapping = aes(x = number_of_rides_given)) +
  geom_density(color="black", fill="lightgray") +
  labs (title = "Distribution of Number of Rides Given", x = "Number of Rides Given Per Driver", y = "Density") +
  theme_minimal() +
  geom_vline(xintercept = max, color="red",
             linetype="dashed") +
  geom_vline(xintercept = second_max, color="red",
             linetype="dashed") +
  annotate("text",x=85,y=0.00375, label = "42 rides", size = 3) +
  annotate("text",x=299,y = 0.00241, label = "252 rides", size = 3)
ggsave(filename = "plots/num_rides_density.png", dpi = 400, width = 7.5, height = 4)
```
